sty working

+++ Assignment_5/graphicsLab/project/data/Zep.frag
@@ -1,6 +1,6 @@
-uniform mediump mat4 ViewMatrix;
-uniform mediump mat4 modelMatrixZep;
-uniform mediump mat4 ProjectionMatrix;
+uniform highp mat4 ViewMatrix;
+uniform highp mat4 modelMatrixZep;
+uniform highp mat4 ProjectionMatrix;
 
 uniform mediump mat3 NormalMatrixZep;
